ingest/requirements.txt
google-genai>=1.0.0
psycopg2-binary>=2.9
pgvector>=0.2.0
python-dotenv>=1.0.0
